feature kill: remove type/value inferring, NWN_LIB_INFER_*
This rather huge commit removes all type/value inferring code
and external support from nwn-lib.

This was broken for quite a while, and is not in the spirit of
nwn-lib - it is experimental code that should never have entered
nwn-lib mainline.

Sorry - if this really inconveniences anyone, please bring this up
with me.