0.4.7-rel
A real bugfix-release, this takes care of a whole swad of fixes
in erf and gff handling, and also improves Windows compatibility.

Particularily, erf writing was kind of broken in the previous release,
putting wrong offsets for files created with nwn-erf.

Please browse the commit log for a full list of changes.