0.4.10-rel
A swath of bugfixes, JSON read/write, and finally .key and .bif reading support.
Also, Resources::Manager got some documentation and a few touches here and there,
as did bin/nwn-erf, which supports partial extraction now.
For our non-ASCII users, there is a ENV variable that can be used to specify the
encoding in which local GFF data is supposed to be in.