rework yaml support for psych & ruby 2.0
This breaks 1.8/1.9 yaml file format compatibility.
taguri was updated to reflect this.

This also now requires all inline strings to be properly UTF-8-encoded,
not with ASCII NWN-style escape sequences. The easiest way to convert
your data files is to first dump them to gff with the old version,
then use the new one one to regenerate the yaml files.